requirements.txt
# CS2 Price Database Requirements
# Core dependencies for the CS2 price collection system

# HTTP client for Steam API requests
aiohttp>=3.8.0

# Environment variable management
python-dotenv>=0.19.0

# Standard library modules (included with Python):
# - asyncio (async/await support)
# - json (JSON parsing)
# - logging (logging functionality) 
# - datetime (date/time handling)
# - typing (type hints)
# - os (operating system interface)
# - re (regular expressions)
# - time (time functions)